move fg assignment and handle single atoms in rdkit
